+++ 전진/250820/triangle_main.java
